moved the data path to paths dir and added auth
Added password authentication and moved the scouting_post handler to paths/data.rs
