+++ src/main.rs
@@ -1,5 +1,5 @@
 use axum::{
-    http::{self, HeaderValue, Method, StatusCode},
+    http::{self, HeaderValue, header::HeaderMap, Method, StatusCode},
     response::{Html, IntoResponse},
     routing::{get, post},
     Json, Router,
@@ -9,6 +9,7 @@ use std::net::SocketAddr;
 use tower_http::cors::CorsLayer;
 mod csvstuff;
 mod settings;
+mod paths;
 use serde_json::Value;
 use std::io::Write;
 
@@ -21,7 +22,7 @@ async fn main() {
 
     let backend = async {
         let app = Router::new()
-            .route("/scouting", post(scouting_post)).layer(
+            .route("/scouting", post(paths::data::scouting_post)).layer(
                 // see https://docs.rs/tower-http/latest/tower_http/cors/index.html
                 // for more details
                 //
@@ -53,48 +54,6 @@ async fn serve(app: Router, port: u16) {
     println!("Running on port: {}", &port);
     success!(format!("Successfully started on port: {}", &port));
     axum::serve(listener, app).await.unwrap();
-}
-
-async fn scouting_post(extract::Json(data): Json<csvstuff::Data>) -> Result<String, StatusCode> {
-    debug_log!("- Scouting data was posted to the server");
-    let settings = settings::Settings::new().unwrap();
-    
-    let mut owned_string: String = "".to_owned(); // String for later to append to
-    let mut thing: String; // Placeholder string
-
-    let mut hash_vec: Vec<(&String, &Value)> = data.data.iter().collect();
-
-    // Sorts the vector by category
-    hash_vec.sort_by(|a, b| {
-        a.1.as_object().expect("Failed to turn into object").get_key_value("category").expect("Failed to get category").1.as_str().expect("Failed to get content").cmp(
-            b.1.as_object().expect("Failed to turn into object").get_key_value("category").expect("Failed to get category").1.as_str().expect("Failed to get content")
-        )
-    });
-
-    // Makes the headers if the file is empty
-    if csvstuff::file_empty(settings.stands_data).unwrap() {
-        success!("File was empty, made headers");
-        let mut header: String = "".to_owned();
-        let mapped: Vec<String> = hash_vec.iter().map(|point| point.0.to_string()).collect();
-        for val in mapped {header.push_str(format!("{},", val).as_str())}
-        let _ = csvstuff::append_csv(&header);
-    }
-
-    for i in hash_vec {
-        // Iterates through the list and appends the data to a string
-        thing = format!("{}, ", i.1.as_object().unwrap().get_key_value("content").expect("Failed to get content").1.to_string().replace(",", ""));
-        owned_string.push_str(&thing)
-    }
-
-    // Adds the information to data.csv
-    match csvstuff::append_csv(&owned_string) {
-        Ok(_e) => {}
-        Err(error) => {
-            error!(format!("Uh oh, {}", error));
-            return Err(StatusCode::INTERNAL_SERVER_ERROR)
-        }
-    } 
-    return Ok("It worked!".into()); // Returns accepted status when done
 }
 
 async fn html() -> impl IntoResponse {
