+++ src/paths/mod.rs
@@ -1,1 +1,2 @@
-pub mod graphs;
+// pub mod graphs;
+pub mod data;
